chore: remove legacy items view file
