+++ inventory/ui_urls.py
@@ -10,22 +10,26 @@ from .views.indents import (
     indent_pdf,
     indent_update_status,
 )
-from .views.items import (
-    ItemCreateHTMXView,
-    ItemCreatePartialView,
+from .views.items.detail import (
     ItemDeleteView,
     ItemDetailView,
     ItemEditView,
     ItemInlineUpdateView,
-    ItemsBulkUpdateView,
-    ItemsBulkUploadView,
+    ItemToggleActiveView,
+)
+from .views.items.list import (
     ItemSearchView,
     ItemsExportView,
     ItemsListView,
     ItemsTableView,
-    ItemToggleActiveView,
-    get_purchase_units,
-    get_subcategories,
+)
+from .views.items.stock import (
+    ItemCreateHTMXView,
+    ItemCreatePartialView,
+    ItemsBulkUpdateView,
+    ItemsBulkUploadView,
+    PurchaseUnitsView,
+    SubcategoriesView,
 )
 from .views.ml import ml_dashboard
 from .views.purchase_orders import (
@@ -81,8 +85,8 @@ urlpatterns = [
     path("items/<int:pk>/", ItemDetailView.as_view(), name="item_detail"),
 
     path("items/search/", ItemSearchView.as_view(), name="item_search"),
-    path("items/purchase-units/", get_purchase_units, name="get_purchase_units"),
-    path("items/subcategories/", get_subcategories, name="get_subcategories"),
+    path("items/purchase-units/", PurchaseUnitsView.as_view(), name="get_purchase_units"),
+    path("items/subcategories/", SubcategoriesView.as_view(), name="get_subcategories"),
     path("items/bulk-upload/", ItemsBulkUploadView.as_view(), name="items_bulk_upload"),
     path("suppliers/", SuppliersListView.as_view(), name="suppliers_list"),
     path("suppliers/table/", SuppliersTableView.as_view(), name="suppliers_table"),
