+++ inventory/views/items/constants.py
@@ -1,0 +1,10 @@
+"""Shared constants for item views."""
+
+EXCLUDED_FIELDS = [
+    "name",
+    "base_unit",
+    "purchase_unit",
+    "category",
+    "sub_category",
+    "departments",
+]  # Fields handled explicitly elsewhere
